+++ resources/js/components/Pages/Dashboard.vue
@@ -3,8 +3,8 @@
 
         <v-toolbar height="48" color="primary" title="Resum" elevation="10">
 
-            <v-btn-toggle v-model="limit" size="small" variant="text" rounded density="compact" divided mandatory>
-                <v-btn :value="v" v-for="v in [1, 3, 5, 10]" :key="v" v-text="v" />
+            <v-btn-toggle v-model="limit" size="small" rounded density="compact" divided mandatory class="mr-2">
+                <v-btn :value="v" v-for="v in [1, 3, 5, 10]" :key="v" v-text="v" :variant="limit == v ? 'elevated' : 'plain'" color="primary" />
             </v-btn-toggle>
         </v-toolbar>
 
@@ -75,37 +75,36 @@
 
 <script setup>
 
-import { useLoaderStore } from '@/stores/loader';
-const loaderStore = useLoaderStore();
+import { useLoaderStore } from '@/stores/loader'
+const loaderStore = useLoaderStore()
 
-import { useDisplay } from 'vuetify';
-const display = useDisplay();
+import { ref, onMounted, onUnmounted, watch } from 'vue'
+import DashboardCard from '../Shared/DashboardCard.vue'
 
-import { ref, onMounted, onUnmounted, watch } from 'vue';
-import DashboardCard from '../Shared/DashboardCard.vue';
-
-const dashboard = ref(false);
-const limit = ref(3);
+const dashboard = ref(false)
+const savedLimit = localStorage.getItem('limit') || 3
+const limit = ref(parseInt(savedLimit))
 var refreshThread
 
 const updateDashboard = () => {
-    loaderStore.add();
+    loaderStore.add()
     fetch(`/api/dashboard?limit=${limit.value}`)
         .then(response => response.json())
         .then(data => dashboard.value = data)
         .catch(error => console.error('Error fetching data:', error))
         .finally(() => {
             loaderStore.remove()
-            refreshThread = setTimeout(() => updateDashboard(), 10000);
-        });
-};
+            refreshThread = setTimeout(() => updateDashboard(), 10000)
+        })
+}
 
-watch(limit, () => {
-    clearTimeout(refreshThread);
+watch(limit, (newValue) => {
+    localStorage.setItem('limit', newValue)
+    clearTimeout(refreshThread)
     updateDashboard()
-})
+});
 
-onMounted(() => updateDashboard());
-onUnmounted(() => clearTimeout(refreshThread));
+onMounted(() => updateDashboard())
+onUnmounted(() => clearTimeout(refreshThread))
 
 </script>