+++ storage/app/public/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 20240912
 
+- Add persistance of dashboard and drivers list modes to localStorage. Improve visuals on those selectors
 - Update Pitskill License systems
 - Update driver refreshing algorithm